requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy
opencv-python
einops
gradio
ninja
packaging
pillow
runpod
